+++ layer.h
@@ -24,14 +24,14 @@
 #include "mymath.h"
 #include <memory>
 
-extern float length_2d(const point_2d &p);
-extern float det_2d(const point_2d &p1, const point_2d &p2);
-extern float dot_2d(const point_2d &p1, const point_2d &p2);
+extern double length_2d(const point_2d &p);
+extern double det_2d(const point_2d &p1, const point_2d &p2);
+extern double dot_2d(const point_2d &p1, const point_2d &p2);
 extern point_2d sub_2d(const point_2d &p1, const point_2d &p2);
 extern point_2d perp_2d(const point_2d &p);
-extern point_2d scale_2d(const point_2d &p, float s);
+extern point_2d scale_2d(const point_2d &p, double s);
 extern bool collide_thick_lines_2d(const point_2d &tl1_p1, const point_2d &tl1_p2,
-	 					const point_2d &tl2_p1, const point_2d &tl2_p2, float r);
+	 					const point_2d &tl2_p1, const point_2d &tl2_p2, double r);
 
 //layer class
 class layer
@@ -51,8 +51,8 @@ public:
 		}
 		point_2d m_p1;
 		point_2d m_p2;
-		float m_radius;
-		float m_gap;
+		double m_radius;
+		double m_gap;
 	};
 
 	struct record
@@ -62,14 +62,14 @@ public:
 			, m_line(l)
 		{
 			auto pv = perp_2d(sub_2d(l.m_p2, l.m_p1));
-			m_lv_norm = scale_2d(pv, 1.0f / length_2d(pv));
+			m_lv_norm = scale_2d(pv, 1.0 / length_2d(pv));
 			m_lv_dist = dot_2d(m_lv_norm, l.m_p1);
 
 			// m_pv_norm = perp_2d(m_lv_norm);
 			// m_pv_dist1 = dot_2d(m_pv_norm, l.m_p1);
 			// m_pv_dist2 = m_pv_dist1 - len;
 		}
-		auto hit(const line &l, float d)
+		auto hit(const line &l, double d)
 		{
 			auto dp1 = dot_2d(m_lv_norm, l.m_p1) - m_lv_dist;
 			auto dp2 = dot_2d(m_lv_norm, l.m_p2) - m_lv_dist;
@@ -86,11 +86,11 @@ public:
 		int m_id;
 		line m_line;
 		point_2d m_lv_norm;
-		float m_lv_dist;
+		double m_lv_dist;
 
 		// point_2d m_pv_norm;
-		// float m_pv_dist1;
-		// float m_pv_dist2;
+		// double m_pv_dist1;
+		// double m_pv_dist2;
 	};
 
 	struct aabb
@@ -104,7 +104,7 @@ public:
 	typedef std::vector<std::shared_ptr<record>> bucket;
 	typedef std::vector<bucket> buckets;
 
-	layer(const dims &dims, float s);
+	layer(const dims &dims, double s);
 	~layer();
 	aabb get_aabb(const line &l);
 	void add_line(const line &l);
@@ -114,7 +114,7 @@ public:
 private:
 	int m_width;
 	int m_height;
-	float m_scale;
+	double m_scale;
 	buckets m_buckets;
 	int m_test;
 };
@@ -134,15 +134,15 @@ public:
 	{
 		point_3d m_p1;
 		point_3d m_p2;
-		float m_radius;
-		float m_gap;
+		double m_radius;
+		double m_gap;
 	};
 
-	layers(const dims &dims, float s);
+	layers(const dims &dims, double s);
 	~layers();
-	void add_line(const point_3d &p1, const point_3d &p2, float r, float g);
-	void sub_line(const point_3d &p1, const point_3d &p2, float r, float g);
-	bool hit_line(const point_3d &p1, const point_3d &p2, float r, float g);
+	void add_line(const point_3d &p1, const point_3d &p2, double r, double g);
+	void sub_line(const point_3d &p1, const point_3d &p2, double r, double g);
+	bool hit_line(const point_3d &p1, const point_3d &p2, double r, double g);
 	void add_line(const line &l);
 	void sub_line(const line &l);
 	bool hit_line(const line &l);
